requirements.txt
PySide6==6.7.2
pdfplumber==0.11.4
pypdf==4.2.0
google-generativeai==0.8.3
